Keep option to choose between carousel and grid
